Refactor to new file and log before unload
